+++ src/main/java/pl/betoncraft/betonquest/notify/ChatNotifyIO.java
@@ -1,7 +1,9 @@
 package pl.betoncraft.betonquest.notify;
 
 import org.bukkit.entity.Player;
+import pl.betoncraft.betonquest.conversation.Conversation;
 import pl.betoncraft.betonquest.exceptions.InstructionParseException;
+import pl.betoncraft.betonquest.utils.PlayerConverter;
 
 import java.util.Map;
 
@@ -14,6 +16,11 @@ public class ChatNotifyIO extends NotifyIO {
 
     @Override
     protected void notifyPlayer(final String message, final Player player) {
-        player.sendMessage(message);
+        final Conversation conversation = Conversation.getConversation(PlayerConverter.getID(player));
+        if (conversation == null || conversation.getInterceptor() == null) {
+            player.sendMessage(message);
+        } else {
+            conversation.getInterceptor().sendMessage(message);
+        }
     }
 }